# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: VAN API
  version: 1.0.0
  description: >
    API REST de VAN para la app móvil (Flutter). Toda la lógica de negocio
    vive en el servidor; la app es un cliente delgado.

    Autenticación: Bearer JWT. El login devuelve access_token (corto,
    30 min) y refresh_token (opaco, 30 días, rotativo). Cuando el
    access_token expira, la app debe llamar POST /auth/refresh con el
    refresh_token para obtener un par nuevo. El refresh_token se revoca
    con POST /auth/logout.

    Envelope de respuestas:
      Exito:  {"success": true,  "data": {...}}
      Error:  {"success": false, "error": {"code": "CODE", "message": "..."}}
...
paths:
...
  /wallet:
    get:
      tags: [wallet]
      summary: Saldo y currency del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          balance: {type: string, example: "1234.50"}
                          currency: {type: string, example: "ARS"}
        '401': { $ref: '#/components/responses/Error' }
  /wallet/transactions:
    get:
      tags: [wallet]
      summary: Movimientos de billetera (paginados, filtro por tipo)
      security:
        - bearerAuth: []
      parameters:
        - {name: type, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
      responses:
        '200':
          description: Lista de transacciones
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: {type: object}
                          pagination: {$ref: '#/components/schemas/Pagination'}
        '401': { $ref: '#/components/responses/Error' }
  /wallet/topups:
    get:
      tags: [wallet]
      summary: Listar recargas propias (paginadas, filtro por status)
      security:
        - bearerAuth: []
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [pending, confirmed, expired]}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: limit, in: query, schema: {type: integer, default: 20, maximum: 100}}
      responses:
        '200':
          description: Lista de topups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: {type: object}
                          pagination: {$ref: '#/components/schemas/Pagination'}
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [wallet]
      summary: Crear recarga de billetera
      description: |
        `mercadopago` → crea preference y devuelve `init_point`.
        `cvu` / `bank` → persiste topup manual (init_point: null).
        Amount 100–500.000.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, method]
              properties:
                amount: {type: number, minimum: 100, maximum: 500000}
                method: {type: string, enum: [mercadopago, cvu, bank]}
      responses:
        '201':
          description: Topup creado
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          topup:
                            type: object
                            properties:
                              id: {type: integer}
                              amount: {type: string}
                              method: {type: string}
                              status: {type: string}
                              admin_note: {type: string, nullable: true}
                              init_point: {type: string, nullable: true}
                              created_at: {type: string, nullable: true}
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /wallet/topups/{id}:
    get:
      tags: [wallet]
      summary: Detalle de una recarga propia
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: Detalle topup
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          topup: {type: object}
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  # ─── A2: Trip status polling ───
...
  /favorites:
    get:
      tags: [auth]
      summary: Rutas frecuentes del pasajero (≥3 usos)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de favoritos
  /wallet/topups/{topup_id}/voucher:
    post:
      tags: [wallet]
      summary: Subir comprobante de transferencia (base64)
      security:
        - bearerAuth: []
      parameters:
        - name: topup_id
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image: {type: string, description: Base64 o data URL}
      responses:
        '200':
          description: Voucher subido
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  # ─── List trips (undocumented) ───
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: {type: boolean, example: false}
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Credenciales inválidas
...
    OkResponse:
      type: object
      properties:
        success: {type: boolean}
        data:
          type: object
          properties:
            ok: {type: boolean}
    Pagination:
      type: object
      properties:
        page: {type: integer}
        limit: {type: integer}
        total: {type: integer}
        pages: {type: integer}
...